Initial commit of example
This is the initial commit of the example site for embedded prototypes. The site has a prototype that can be manipulated using controls on the site. The site also includes an event viewer to see prettified JSON versions of the events emitted by the prototype. Additionally, the buttons on the site are enabled and disabled based on events sent by the prototype.